src/services/route/CarRouteOptimizer.ts
import { MCPIntegrationService } from '@/services/integration/MCPIntegrationService';
import {
  GeoPoint,
  RoutePreferences,
  OptimizationResult,
  RouteMetrics,
  TerrainDifficulty,
  TerrainConditions
} from '@/types/route/types';
import { WeatherConditions } from '@/types/weather';
import logger from '@/utils/logger';

export class CarRouteOptimizer {
  constructor(
    private readonly mcpService: MCPIntegrationService
  ) {}

  async optimizeRoute(
    startPoint: GeoPoint,
    endPoint: GeoPoint,
    preferences: RoutePreferences
  ): Promise<OptimizationResult> {
    try {
      // Get car-specific route data
      const trafficData = await this.mcpService.getTrafficData(startPoint, endPoint);
      const roadConditions = await this.mcpService.getRoadConditions(startPoint, endPoint);
      const restrictions = await this.mcpService.getRoadRestrictions(startPoint, endPoint);
      const elevationProfile = await this.mcpService.getElevationProfile(startPoint, endPoint);

      // Calculate optimal route considering car-specific factors
      const baseRoute = await this.mcpService.getBaseRoute(startPoint, endPoint);
      
      // Apply car-specific optimizations
      let optimizedPath = this.optimizePath(
        baseRoute.path,
        trafficData,
        roadConditions,
        restrictions,
        preferences
      );

      // Calculate metrics
      const { distance, duration } = this.calculateBaseMetrics(optimizedPath);
      const elevation = this.calculateElevation(optimizedPath, elevationProfile);
      const { fuelEfficiency, trafficDelay, stopCount } = this.calculateCarSpecificMetrics(
        optimizedPath,
        trafficData,
        roadConditions
      );

      // Generate warnings
      const warnings = this.generateWarnings(trafficData, roadConditions, restrictions);

      return {
        path: optimizedPath,
        metrics: {
          distance,
          duration,
          elevation,
          speed: distance / duration,
          fuelEfficiency,
          trafficDelay,
          stopCount,
          safety: this.calculateSafetyScore(roadConditions, trafficData),
          weatherImpact: null,
          terrainDifficulty: TerrainDifficulty.MODERATE,
          surfaceType: 'paved'
        },
        warnings
      };
    } catch (error) {
      logger.error('Failed to optimize car route:', error);
      throw error;
    }
  }

  private optimizePath(
    basePath: GeoPoint[],
    trafficData: any,
    roadConditions: any,
    restrictions: any,
    preferences: RoutePreferences
  ): GeoPoint[] {
    // Implement path optimization logic
    return basePath;
  }

  private calculateBaseMetrics(path: GeoPoint[]): { distance: number; duration: number } {
    // Implement base metrics calculation
    return {
      distance: 1000,
      duration: 600
    };
  }

  private calculateElevation(path: GeoPoint[], elevationProfile: any): { gain: number; loss: number; profile: number[] } {
    // Implement elevation calculation
    return {
      gain: 100,
      loss: 50,
      profile: []
    };
  }

  private calculateCarSpecificMetrics(
    path: GeoPoint[],
    trafficData: any,
    roadConditions: any
  ): { fuelEfficiency: number; trafficDelay: number; stopCount: number } {
    // Implement car-specific metrics calculation
    return {
      fuelEfficiency: 8.5,
      trafficDelay: 120,
      stopCount: 4
    };
  }

  private calculateSafetyScore(roadConditions: any, trafficData: any): number {
    // Implement safety score calculation
    return 0.85;
  }

  private generateWarnings(
    trafficData: any,
    roadConditions: any,
    restrictions: any
  ): string[] {
    const warnings: string[] = [];

    // Add traffic warnings
    const congestionPoints = trafficData.segments
      .filter(segment => segment.congestionLevel > 0.7)
      .map(segment => ({
        start: segment.start,
        end: segment.end
      }));

    if (congestionPoints.length > 0) {
      warnings.push('Heavy traffic conditions detected');
    }

    // Add road condition warnings
    const poorConditionSegments = roadConditions.segments
      .filter(segment => segment.condition === 'poor')
      .map(segment => ({
        start: segment.start,
        end: segment.end
      }));

    if (poorConditionSegments.length > 0) {
      warnings.push('Poor road conditions detected');
    }

    // Add restriction warnings
    const restrictedSegments = restrictions.segments
      .filter(segment => segment.restricted)
      .map(segment => ({
        start: segment.start,
        end: segment.end
      }));

    if (restrictedSegments.length > 0) {
      warnings.push('Road restrictions in effect');
    }

    return warnings;
  }

  private async calculateAlternativeRoutes(
    startPoint: GeoPoint,
    endPoint: GeoPoint,
    preferences: RoutePreferences
  ): Promise<number> {
    const alternatives = await this.mcpService.getAlternativeRoutes(
      startPoint,
      endPoint,
      preferences
    );
    return alternatives.routes.length;
  }
} 